refactor: route couch replication through httpClientApi, rename to useFetchCouch
The dedicated HttpCouchClient instance registered the same injectToken +
refreshOnUnauthorized interceptors as httpClientApi, so it added no behavior.
Route useFetchCouch through the shared client and rename the file to match
its only remaining export.
